MT: spatula-ize
